gc_worker: Split config and gc manager out to make the file smaller
